fix(web): extract Auth0 returnTo capture into separate module (tc-0rt0)
Auth0AuthAdapter.tsx exported both a component and a non-component
function (captureAuth0RedirectReturnTo), tripping eslint's
react-refresh/only-export-components rule and failing PR #363's gate.

Move the capture function and its module-level state into
auth0-redirect-return-to.ts. Auth0AuthAdapter now reads the pending
value via readPendingAuth0RedirectReturnTo() and only exports the
component. App.tsx and the adapter test import the capture from the
new module.
