+++ web/src/App.tsx
@@ -2,7 +2,8 @@ import { BrowserRouter } from 'react-router-dom';
 import { Auth0Provider, type AppState } from '@auth0/auth0-react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { loadAuthConfig } from './auth/auth-config';
-import { Auth0AuthAdapter, captureAuth0RedirectReturnTo } from './auth/Auth0AuthAdapter';
+import { Auth0AuthAdapter } from './auth/Auth0AuthAdapter';
+import { captureAuth0RedirectReturnTo } from './auth/auth0-redirect-return-to';
 import { ApiClientProvider } from './api/ApiClientProvider';
 import { ProfileRepositoryProvider } from './auth/ProfileRepositoryProvider';
 import { AppRoutes } from './AppRoutes';
