+++ web/src/auth/Auth0AuthAdapter.tsx
@@ -2,22 +2,8 @@ import type { ReactNode } from 'react';
 import { useMemo, useCallback, useState } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import { AuthProvider } from './auth-context';
+import { readPendingAuth0RedirectReturnTo } from './auth0-redirect-return-to';
 import type { AuthPort, LoginWithRedirectOptions } from '../domain/ports/auth-port';
-
-/**
- * Module-level holder for the `returnTo` path captured from `appState.returnTo`
- * by Auth0's `onRedirectCallback`. The Auth0 React SDK only exposes `appState`
- * via that callback prop on `Auth0Provider` (not via `useAuth0()`), so we stash
- * it here and let `Auth0AuthAdapter` surface it through `AuthPort.returnTo`.
- *
- * Wire `onRedirectCallback={(appState) => captureAuth0RedirectReturnTo(appState?.returnTo)}`
- * on the `Auth0Provider` in the composition root.
- */
-let pendingReturnTo: string | undefined = undefined;
-
-export function captureAuth0RedirectReturnTo(value: string | undefined): void {
-  pendingReturnTo = value;
-}
 
 interface Props {
   children: ReactNode;
@@ -26,10 +12,7 @@ interface Props {
 export function Auth0AuthAdapter({ children }: Props) {
   const { isAuthenticated, isLoading, loginWithRedirect, logout: auth0Logout, error } = useAuth0();
 
-  // Snapshot the captured returnTo at mount so consumers see a stable value.
-  // The capture itself happens before React mounts the auth subtree, when
-  // Auth0Provider invokes onRedirectCallback during the post-callback bootstrap.
-  const [returnTo] = useState<string | undefined>(() => pendingReturnTo);
+  const [returnTo] = useState<string | undefined>(() => readPendingAuth0RedirectReturnTo());
 
   const logout = useCallback(
     () =>
