+++ web/src/auth/auth0-redirect-return-to.ts
@@ -1,0 +1,9 @@
+let pendingReturnTo: string | undefined = undefined;
+
+export function captureAuth0RedirectReturnTo(value: string | undefined): void {
+  pendingReturnTo = value;
+}
+
+export function readPendingAuth0RedirectReturnTo(): string | undefined {
+  return pendingReturnTo;
+}
